Tests updated and assured all tests pass

+++ tests/accessRestrictionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once '../vendor/autoload.php';
+require_once 'init.php';
 
 use acdhOeaw\fedora\Fedora;
 use acdhOeaw\fedora\acl\WebAclRule as AR;
@@ -8,14 +8,12 @@ use acdhOeaw\util\Indexer;
 use acdhOeaw\util\RepoConfig as RC;
 use acdhOeaw\util\ResourceFactory as RF;
 
-RC::init('../config.ini');
-RC::set('sparqlUrl', 'https://fedora.localhost/blazegraph/sparql');
 RC::set('containerDir', __DIR__ . '/..');
 RC::set('containerToUriPrefix', 'test://');
 RC::set('indexerDefaultBinaryClass', 'https://vocabs.acdh.oeaw.ac.at/schema#Resource');
 RC::set('indexerDefaultCollectionClass', 'https://vocabs.acdh.oeaw.ac.at/schema#Collection');
 $rightsProp = RC::get('fedoraAccessRestrictionProp');
-$roleProp = RC::get('fedoraAccessRoleProp');
+$roleProp   = RC::get('fedoraAccessRoleProp');
 $fedora     = new Fedora();
 RF::init($fedora);
 
@@ -28,10 +26,13 @@ try {
     assert(AR::READ === $res->getAcl(true)->getMode(AR::USER, AR::PUBLIC_USER));
 } catch (Exception $e) {
     $fedora->rollback();
+    throw $e;
 } finally {
-    $fedora->begin();
-    $res->delete(true, true, true);
-    $fedora->commit();
+    if (isset($res)) {
+        $fedora->begin();
+        $res->delete(true, true, true);
+        $fedora->commit();
+    }
 }
 
 echo "Academic access restriction is properly granted\n";
@@ -45,10 +46,13 @@ try {
     assert(AR::READ === $acl->getMode(AR::USER, RC::get('academicGroup')));
 } catch (Exception $e) {
     $fedora->rollback();
+    throw $e;
 } finally {
-    $fedora->begin();
-    $res->delete(true, true, true);
-    $fedora->commit();
+    if (isset($res)) {
+        $fedora->begin();
+        $res->delete(true, true, true);
+        $fedora->commit();
+    }
 }
 
 echo "Restricted access restriction is properly granted\n";
@@ -62,10 +66,13 @@ try {
     assert(AR::NONE === $acl->getMode(AR::USER, RC::get('academicGroup')));
 } catch (Exception $e) {
     $fedora->rollback();
+    throw $e;
 } finally {
-    $fedora->begin();
-    $res->delete(true, true, true);
-    $fedora->commit();
+    if (isset($res)) {
+        $fedora->begin();
+        $res->delete(true, true, true);
+        $fedora->commit();
+    }
 }
 
 echo "Escalating access rights works\n";
@@ -90,10 +97,13 @@ try {
     assert(AR::READ === $acl->getMode(AR::USER, AR::PUBLIC_USER));
 } catch (Exception $e) {
     $fedora->rollback();
+    throw $e;
 } finally {
-    $fedora->begin();
-    $res->delete(true, true, true);
-    $fedora->commit();
+    if (isset($res)) {
+        $fedora->begin();
+        $res->delete(true, true, true);
+        $fedora->commit();
+    }
 }
 
 echo "Limiting access rights works\n";
@@ -118,19 +128,22 @@ try {
     assert(AR::NONE === $acl->getMode(AR::USER, RC::get('academicGroup')));
 } catch (Exception $e) {
     $fedora->rollback();
+    throw $e;
 } finally {
-    $fedora->begin();
-    $res->delete(true, true, true);
-    $fedora->commit();
+    if (isset($res)) {
+        $fedora->begin();
+        $res->delete(true, true, true);
+        $fedora->commit();
+    }
 }
 
 echo "hasAccessRole works for restricted resources\n";
 try {
     $fedora->begin();
     $res = RF::create([
-        'type' => RC::get('fedoraRepoObjectClass'), 
-        $rightsProp => 'restricted',
-        $roleProp => 'testUser'
+            'type'      => RC::get('fedoraRepoObjectClass'),
+            $rightsProp => 'restricted',
+            $roleProp   => 'testUser'
     ]);
     $fedora->commit();
 
@@ -138,17 +151,24 @@ try {
     assert(AR::READ === $acl->getMode(AR::USER, 'testUser'));
 } catch (Exception $e) {
     $fedora->rollback();
+    throw $e;
 } finally {
-    $fedora->begin();
-    $res->delete(true, true, true);
-    $fedora->commit();
+    if (isset($res)) {
+        $fedora->begin();
+        $res->delete(true, true, true);
+        $fedora->commit();
+    }
 }
 
 echo "Larger import works\n";
 try {
+    $resources = [];
+
     $fedora->begin();
     $coll      = RF::create(['type' => RC::get('fedoraRepoObjectClass')]);
     $ind       = new Indexer($coll);
+    $ind->setBinaryClass(RC::get('fedoraRepoObjectClass'));
+    $ind->setCollectionClass(RC::get('fedoraRepoObjectClass'));
     $ind->setPaths(['src']);
     $resources = $ind->index();
     $fedora->commit();
@@ -158,11 +178,11 @@ try {
     }
 } catch (Exception $e) {
     $fedora->rollback();
+    throw $e;
 } finally {
-    $fedora->begin();
-    foreach ($resources as $res) {
-        $res->delete(true, true, true);
+    if (isset($coll)) {
+        $fedora->begin();
+        $coll->delete(true, true, true);
+        $fedora->commit();
     }
-    $coll->delete(true, true, true);
-    $fedora->commit();
 }
